Revert "feat: implement useMutateApprovalDecision hook "
